+++ src/Waves.Core/GameContext/KuroGameContextBase/Commons/DownloadZip.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Waves.Core.GameContext;
+
+/// <summary>
+/// Zip更新下载
+/// </summary>
+partial class KuroGameContextBase
+{
+
+}
