+++ src/Waves.Core/GameContext/KuroGameContextBase/KuroGameContextBase.Downloader.cs
@@ -20,48 +20,7 @@ namespace Waves.Core.GameContext;
 
 public partial class KuroGameContextBase
 {
-    /// <summary>
-    /// 下载校验最大并发数
-    /// </summary>
-    const int MAX_Concurrency_Count = 6;
-
-    #region 常量
-    const int MaxBufferSize = 65536; // 64KB缓冲区
-    const long UpdateThreshold = 1048576; // 1MB进度更新阈值
-    #endregion
-
-    #region 字段和属性
-    private string _downloadBaseUrl;
-    private long _totalfileSize = 0L;
-    private long _totalProgressSize = 0L;
-    private long _totalFileTotal = 0L;
-    private long _totalProgressTotal = 0L;
-    string baseUrl = "";
-    #endregion
-
-    #region DownloadStatus
-    private long _totalVerifiedBytes;
-    private long _totalDownloadedBytes;
-    private DateTime _lastSpeedUpdateTime;
-    private double _downloadSpeed;
-    private double _verifySpeed;
-
-    private GameContextOutputArgs? _lastOutputArgs;
-    private GameContextOutputArgs? _lastProdOutputArgs;
-
-    private DateTime _lastSpeedTime = DateTime.Now;
-    private long _lastSpeedBytes; // 速度计算基准值
-    #endregion
-
-    #region 速度属性
-    public double DownloadSpeed => _downloadSpeed;
-    public double VerifySpeed => _verifySpeed;
-    #endregion
-    #region DownloadStatus
-    private DownloadState _downloadState;
-    private DownloadState _prodDownloadState;
-    private CancellationTokenSource _prodDownloadCTS;
-    #endregion
+    
     #region 公开方法
     public async Task StartDownloadTaskAsync(
         string folder,
@@ -285,21 +244,6 @@ public partial class KuroGameContextBase
         _downloadState.IsActive = false;
     }
 
-    public int GetVerifyLimit()
-    {
-        if (
-            this.ContextName == nameof(PunishMainGameContext)
-            || this.ContextName == nameof(PunishGlobalGameContext)
-            || this.ContextName == nameof(PunishTwGameContext)
-        )
-        {
-            return 200;
-        }
-        else
-        {
-            return 2000;
-        }
-    }
 
     public async Task<bool> ParallelDownloadAsync(
         List<IndexResource> resource,
@@ -546,12 +490,10 @@ public partial class KuroGameContextBase
         var currentVersion = await GameLocalConfig.GetConfigAsync(
             GameLocalSettingName.LocalGameVersion
         );
-        #region 测试预下载
+
         var previous = launcher
             .ResourceDefault.Config.PatchConfig.Where(x => x.Version == currentVersion)
             .FirstOrDefault();
-
-        #endregion
         PatchIndexGameResource? patch = null;
         this._downloadState = new DownloadState();
         await _downloadState.SetSpeedLimitAsync(this.SpeedValue);
@@ -799,62 +741,6 @@ public partial class KuroGameContextBase
         return launcherIndex;
     }
 
-    private async Task<bool> CheckApplyFilesMd5(
-        List<IndexResource> list,
-        string folder,
-        string tempFolder,
-        Dictionary<string, string> newFiles
-    )
-    {
-        try
-        {
-            var keys = newFiles.Keys.ToList();
-            for (int i = 0; i < keys.Count; i++)
-            {
-                string key = keys[i];
-                string value = newFiles[key];
-                if (File.Exists(value))
-                    File.Delete(value);
-                File.Move(key, value);
-                this.gameContextOutputDelegate?.Invoke(
-                        this,
-                        new GameContextOutputArgs()
-                        {
-                            Type = GameContextActionType.DeleteFile,
-                            FileTotal = keys.Count,
-                            CurrentFile = i,
-                            DeleteString = $"正在移动校验文件{System.IO.Path.GetFileName(value)}",
-                        }
-                    )
-                    .ConfigureAwait(false);
-            }
-            await InitializeProgress(list);
-            var resource = await this.GetGameLauncherSourceAsync();
-            var resourceOne = await this.GetGameResourceAsync(resource.ResourceDefault);
-            if (!await GetGameResourceAsync(folder, resource, false))
-            {
-                await UpdateFileProgress(
-                        GameContextActionType.TipMessage,
-                        0,
-                        false,
-                        false,
-                        "更新校验出错，请直接尝试修复游戏，下载缓存需手动删除"
-                    )
-                    .ConfigureAwait(false);
-                await SetNoneStatusAsync();
-                return false;
-            }
-            return true;
-        }
-        catch (Exception ex)
-        {
-            await SetNoneStatusAsync().ConfigureAwait(false);
-            await UpdateFileProgress(GameContextActionType.TipMessage, 0, false, false, ex.Message);
-            this._isDownload = false;
-            Logger.WriteError(ex.Message);
-            return false;
-        }
-    }
 
     private async Task<bool> DownloadGroupPatcheToResource(
         GameLauncherSource resource,
@@ -912,50 +798,6 @@ public partial class KuroGameContextBase
         return true;
     }
 
-    private async Task<int> DecompressKrdiffFile(
-        string folder,
-        string? krdiffPath,
-        int curent,
-        int total,
-        string? tempFolder = null
-    )
-    {
-        if (krdiffPath == null)
-            return -1000;
-        DiffDecompressManager manager = new DiffDecompressManager(
-            folder,
-            tempFolder ?? folder,
-            krdiffPath
-        );
-        IProgress<(double, double)> progress = new Progress<(double, double)>();
-        ((Progress<(double, double)>)progress).ProgressChanged += async (s, e) =>
-        {
-            if (gameContextOutputDelegate == null)
-                return;
-            await gameContextOutputDelegate
-                .Invoke(
-                    this,
-                    new GameContextOutputArgs
-                    {
-                        Type = GameContextActionType.Decompress,
-                        CurrentSize = (long)e.Item1,
-                        TotalSize = (long)e.Item2,
-                        DownloadSpeed = 0,
-                        VerifySpeed = 0,
-                        RemainingTime = TimeSpan.FromMicroseconds(0),
-                        IsAction = _downloadState?.IsActive ?? false,
-                        IsPause = _downloadState?.IsPaused ?? false,
-                        TipMessage = "正在解压合并资源",
-                        CurrentDecompressCount = curent,
-                        MaxDecompressValue = total,
-                    }
-                )
-                .ConfigureAwait(false);
-        };
-        var result = await manager.StartAsync(progress);
-        Logger.WriteInfo($"解压程序结果{result}");
-        return result;
-    }
 
     private async Task<bool> UpdateGameToResources(string folder, List<IndexResource> resource)
     {
@@ -1122,383 +964,15 @@ public partial class KuroGameContextBase
             _downloadCTS.Dispose();
             _downloadCTS = null;
         }
-        return;
-    }
-
-    #region 校验逻辑
-    /// <summary>
-    /// 校验
-    /// </summary>
-    /// <param name="file"></param>
-    /// <param name="filePath"></param>
-    /// <returns></returns>
-    private async Task<bool> ValidateFileChunks(
-        IndexChunkInfo file,
-        string filePath,
-        bool isPred = false
-    )
-    {
-        using (
-            var fs = new FileStream(
-                filePath,
-                FileMode.Open,
-                FileAccess.Read,
-                FileShare.Read,
-                262144,
-                true
-            )
-        )
+        if (_prodDownloadCTS != null)
         {
-            try
-            {
-                //if (fs.Length < file.End + 1) // 检查文件长度是否足够
-                //{
-                //    Debug.WriteLine($"文件长度不足: {fs.Length} < {file.End + 1}");
-                //    return true;
-                //}
-                var memoryPool = ArrayPool<byte>.Shared;
-                    var downloadCts = GetCTS(isPred);
-                    var state = GetState(isPred);
-                    if (downloadCts == null || state?.IsStop == true)
-                {
-                    throw new OperationCanceledException();
-                }
-                long offset = file.Start;
-                long remaining = file.End - file.Start + 1;
-                bool isValid = true;
-                fs.Seek(offset, SeekOrigin.Begin);
-                using (var md5 = MD5.Create())
-                {
-                    long accumulatedBytes = 0L;
-                    while (remaining > 0 && isValid)
-                    {
-                        if (state != null)
-                            await state.PauseToken.WaitIfPausedAsync();
-                        var buffer = memoryPool.Rent(MaxBufferSize);
-                        try
-                        {
-                            if (
-                                downloadCts.IsCancellationRequested
-                                || state?.IsStop == true
-                            )
-                            {
-                                throw new OperationCanceledException();
-                            }
-                            int bytesRead = await fs.ReadAsync(
-                                    buffer,
-                                    0,
-                                    MaxBufferSize,
-                                    downloadCts.Token
-                                )
-                                .ConfigureAwait(false);
-                            if (bytesRead == 0)
-                            {
-                                break;
-                            }
-                            md5.TransformBlock(buffer, 0, bytesRead, null, 0);
-                            remaining -= bytesRead;
-                            accumulatedBytes += bytesRead;
-                            if (accumulatedBytes >= UpdateThreshold)
-                            {
-                                await UpdateFileProgress(
-                                        GameContextActionType.Verify,
-                                        accumulatedBytes,
-                                        false,
-                                        isPred
-                                    )
-                                    .ConfigureAwait(false);
-                                accumulatedBytes = 0;
-                            }
-                        }
-                        catch (IOException ex)
-                        {
-                            Logger.WriteError(ex.Message);
-                        }
-                        finally
-                        {
-                            memoryPool.Return(buffer);
-                        }
-                    }
-                    if (accumulatedBytes > 0 && accumulatedBytes < UpdateThreshold)
-                    {
-                        await UpdateFileProgress(
-                                GameContextActionType.Verify,
-                                accumulatedBytes,
-                                false,
-                                isPred
-                            )
-                            .ConfigureAwait(false);
-                    }
-                    md5.TransformFinalBlock(Array.Empty<byte>(), 0, 0);
-                    string hash = BitConverter.ToString(md5.Hash!).Replace("-", "").ToLower();
-                    isValid = hash == file.Md5.ToLower();
-                    Logger.WriteInfo($"分片校验结果{hash}|{file.Md5}");
-                    return !isValid;
-                }
-            }
-            catch (IOException ex)
-            {
-                Logger.WriteError(ex.Message);
-                return false;
-            }
-            catch (OperationCanceledException)
-            {
-                throw new OperationCanceledException();
-            }
-            finally
-            {
-                fs.Close();
-                fs.Dispose();
-            }
+            await _prodDownloadCTS.CancelAsync();
+            _prodDownloadCTS.Dispose();
+            _prodDownloadCTS = null;
         }
     }
-
-    private async Task<bool> VaildateFullFile(string md5Value, string filePath, bool isPred = false)
-    {
-        const int bufferSize = 262144; // 80KB缓冲区
-        using var md5 = MD5.Create();
-        var memoryPool = ArrayPool<byte>.Shared;
-        var downloadCts = GetCTS(isPred);
-        var state = GetState(isPred);
-                if (downloadCts == null || state?.IsStop == true)
-        {
-            throw new OperationCanceledException();
-        }
-        const long UpdateThreshold = 1048576; // 1MB进度更新阈值
-        FileStream? fs = null;
-        try
-        {
-            using (
-                fs = new FileStream(
-                    filePath,
-                    FileMode.Open,
-                    FileAccess.Read,
-                    FileShare.Read,
-                    bufferSize: bufferSize,
-                    true
-                )
-            )
-            {
-                bool isBreak = false;
-                long accumulatedBytes = 0L;
-                while (true)
-                {
-                    if (downloadCts.IsCancellationRequested || state?.IsStop == true)
-                    {
-                        throw new OperationCanceledException();
-                    }
-                    //暂停锁
-                    if (state != null)
-                        await state.PauseToken.WaitIfPausedAsync().ConfigureAwait(false);
-                    byte[] buffer = memoryPool.Rent(bufferSize);
-                    try
-                    {
-                        int bytesRead = await fs.ReadAsync(
-                                buffer.AsMemory(0, bufferSize),
-                                downloadCts.Token
-                            )
-                            .ConfigureAwait(false);
-                        if (bytesRead == 0)
-                        {
-                            isBreak = true;
-                            break;
-                        }
-                        md5.TransformBlock(buffer, 0, bytesRead, null, 0);
-                        accumulatedBytes += bytesRead; // 添加此行以累加字节数
-                        if (accumulatedBytes >= UpdateThreshold)
-                        {
-                            await UpdateFileProgress(
-                                    GameContextActionType.Verify,
-                                    accumulatedBytes,
-                                    false,
-                                    isPred
-                                )
-                                .ConfigureAwait(false);
-                            accumulatedBytes = 0;
-                        }
-                    }
-                    finally
-                    {
-                        memoryPool.Return(buffer);
-                    }
-                }
-                if (accumulatedBytes < UpdateThreshold)
-                {
-                    await UpdateFileProgress(
-                            GameContextActionType.Verify,
-                            accumulatedBytes,
-                            false,
-                            isPred: isPred
-                        )
-                        .ConfigureAwait(false);
-                }
-            }
-
-            md5.TransformFinalBlock(Array.Empty<byte>(), 0, 0);
-            string hash = BitConverter.ToString(md5.Hash!).Replace("-", "").ToLower();
-
-            return !(hash == md5Value);
-        }
-        catch (OperationCanceledException)
-        {
-            throw new OperationCanceledException();
-        }
-        catch (IOException ex)
-        {
-            Logger.WriteError(ex.Message);
-            return false;
-        }
-        catch (Exception ex)
-        {
-            Logger.WriteError(ex.Message);
-            return false;
-        }
-        finally
-        {
-            fs?.Close();
-            fs?.Dispose();
-        }
-    }
-
-    #endregion
 
     #region 下载逻辑
-    private async Task DownloadFileByChunks(
-        string dest,
-        string filePath,
-        long start,
-        long end,
-        bool isLast = false,
-        long allSize = 0L,
-        bool isPred = false
-    )
-    {
-        using (
-            var fileStream = new FileStream(
-                filePath,
-                FileMode.OpenOrCreate,
-                FileAccess.ReadWrite,
-                FileShare.Read,
-                262144,
-                true
-            )
-        )
-        {
-            try
-            {
-                var downloadCts = GetCTS(isPred);
-                var state = GetState(isPred);
-                if (downloadCts == null || state?.IsStop == true)
-                {
-                    throw new OperationCanceledException();
-                }
-                long accumulatedBytes = 0;
-                if (start == 0 && end == -1)
-                {
-                    Logger.WriteError($"文件{filePath}，分片数据错误，start={start},end={end}");
-                }
-                using var request = new HttpRequestMessage(
-                    HttpMethod.Get,
-                    _downloadBaseUrl.TrimEnd('/') + "/" + dest.TrimStart('/')
-                );
-                request.Headers.Range = new RangeHeaderValue(start, end);
-                using var response = await HttpClientService.GameDownloadClient.SendAsync(
-                    request,
-                    HttpCompletionOption.ResponseHeadersRead,
-                    downloadCts.Token
-                );
-                var stream = await response.Content.ReadAsStreamAsync(downloadCts.Token);
-                if (start < 0 || end < start)
-                {
-                    Logger.WriteError($"分片范围无效: {start}-{end}");
-                    throw new ArgumentException($"分片范围无效: {start}-{end}");
-                }
-
-                long totalWritten = 0;
-                long chunkTotalSize = end - start + 1;
-                var memoryPool = ArrayPool<byte>.Shared;
-                fileStream.Seek(start, SeekOrigin.Begin);
-                bool isBreak = false;
-                while (totalWritten < chunkTotalSize)
-                {
-                    if (downloadCts.IsCancellationRequested || state?.IsStop == true)
-                    {
-                        throw new OperationCanceledException();
-                    }
-                    if (state != null)
-                        await state.PauseToken.WaitIfPausedAsync().ConfigureAwait(false); // 暂停检查也异步化
-                    int bytesToRead = (int)Math.Min(MaxBufferSize, chunkTotalSize - totalWritten);
-                    byte[] buffer = ArrayPool<byte>.Shared.Rent(bytesToRead);
-                    try
-                    {
-                        int bytesRead = await stream
-                            .ReadAsync(buffer.AsMemory(0, bytesToRead), downloadCts.Token)
-                            .ConfigureAwait(false);
-                        if (bytesRead == 0)
-                        {
-                            isBreak = true;
-                            break;
-                        }
-                    if (state != null)
-                        await state
-                            .SpeedLimiter.LimitAsync(bytesRead)
-                            .ConfigureAwait(false);
-                        await fileStream
-                            .WriteAsync(buffer.AsMemory(0, bytesRead), downloadCts.Token)
-                            .ConfigureAwait(false);
-                        totalWritten += bytesRead;
-                        accumulatedBytes += bytesRead;
-                        if (accumulatedBytes >= UpdateThreshold)
-                        {
-                            await UpdateFileProgress(
-                                    GameContextActionType.Download,
-                                    accumulatedBytes,
-                                    true,
-                                    isPred
-                                )
-                                .ConfigureAwait(false);
-                            accumulatedBytes = 0;
-                        }
-                    }
-                    catch (OperationCanceledException)
-                    {
-                        throw;
-                    }
-                    catch (Exception ex)
-                    {
-                        Logger.WriteError(ex.Message);
-                    }
-                    finally
-                    {
-                        ArrayPool<byte>.Shared.Return(buffer);
-                    }
-                }
-                if (accumulatedBytes > 0 && !isBreak)
-                {
-                    await UpdateFileProgress(
-                            GameContextActionType.Download,
-                            accumulatedBytes,
-                            isPred: isPred
-                        )
-                        .ConfigureAwait(false);
-                }
-                if (isLast)
-                    fileStream.SetLength(allSize);
-                stream.Close();
-                await stream.DisposeAsync();
-            }
-            catch (OperationCanceledException)
-            {
-                throw;
-            }
-            catch (Exception)
-            {
-                await fileStream.FlushAsync(_downloadCTS?.Token ?? CancellationToken.None);
-                await fileStream.FlushAsync();
-                await fileStream.DisposeAsync();
-            }
-        }
-    }
 
     public async Task SetNoneStatusAsync(bool isPred = false)
     {
@@ -1547,130 +1021,6 @@ public partial class KuroGameContextBase
             GameLocalSettingName.LimitSpeed,
             bytesPerSecond.ToString()
         );
-    }
-
-    private async Task DownloadFileByFull(
-        string dest,
-        long size,
-        string filePath,
-        IndexChunkInfo chunk,
-        bool isPred = false
-    )
-    {
-        long accumulatedBytes = 0;
-        using (
-            var fileStream = new FileStream(
-                filePath,
-                FileMode.OpenOrCreate,
-                FileAccess.ReadWrite,
-                FileShare.None,
-                262144,
-                true
-            )
-        )
-        {
-            try
-            {
-                if (chunk.Start == 0 && chunk.End == -1)
-                {
-                    Logger.WriteError(
-                        $"文件{filePath}，分片数据错误，start={chunk.Start},end={chunk.End}"
-                    );
-                    return;
-                }
-                using var request = new HttpRequestMessage(
-                    HttpMethod.Get,
-                    _downloadBaseUrl.TrimEnd('/') + "/" + dest.TrimStart('/')
-                );
-                request.Headers.Range = new RangeHeaderValue(chunk.Start, chunk.End);
-                var downloadCts = GetCTS(isPred);
-                var state = GetState(isPred);
-                using var response = await HttpClientService
-                    .GameDownloadClient.SendAsync(
-                        request,
-                        HttpCompletionOption.ResponseHeadersRead,
-                        downloadCts.Token
-                    )
-                    .ConfigureAwait(false); // 非UI上下文切换
-
-                response.EnsureSuccessStatusCode();
-                var stream = await response
-                    .Content.ReadAsStreamAsync(downloadCts.Token)
-                    .ConfigureAwait(false);
-                if (chunk.Start < 0 || chunk.End < chunk.Start)
-                {
-                    Logger.WriteError($"分片范围无效，start={chunk.Start},end={chunk.End}");
-                    throw new ArgumentException($"分片范围无效: {chunk.Start}-{chunk.End}");
-                }
-
-                long totalWritten = 0;
-                long chunkTotalSize = chunk.End - chunk.Start + 1;
-                var memoryPool = ArrayPool<byte>.Shared;
-
-                fileStream.Seek(chunk.Start, SeekOrigin.Begin);
-                bool isBreak = false;
-                while (totalWritten < chunkTotalSize)
-                {
-                    if (downloadCts.IsCancellationRequested || state?.IsStop == true)
-                    {
-                        throw new OperationCanceledException();
-                    }
-                    if (state != null)
-                        await state.PauseToken.WaitIfPausedAsync().ConfigureAwait(false); // 暂停检查也异步化
-                    int bytesToRead = (int)Math.Min(MaxBufferSize, chunkTotalSize - totalWritten);
-                    byte[] buffer = memoryPool.Rent(bytesToRead);
-                    int bytesRead = await stream
-                        .ReadAsync(buffer.AsMemory(0, bytesToRead), downloadCts.Token)
-                        .ConfigureAwait(false);
-                    if (bytesRead == 0)
-                    {
-                        isBreak = true;
-                    }
-                    if (state != null)
-                        await state.SpeedLimiter.LimitAsync(bytesRead).ConfigureAwait(false);
-                    await fileStream
-                        .WriteAsync(buffer.AsMemory(0, bytesRead), downloadCts.Token)
-                        .ConfigureAwait(false);
-                    totalWritten += bytesRead;
-                    accumulatedBytes += bytesRead;
-                    if (accumulatedBytes >= UpdateThreshold)
-                    {
-                        await UpdateFileProgress(
-                                GameContextActionType.Download,
-                                accumulatedBytes,
-                                true,
-                                isPred
-                            )
-                            .ConfigureAwait(false);
-                        accumulatedBytes = 0; // 重置累积计数器
-                    }
-                }
-                if (accumulatedBytes > 0 && !isBreak)
-                {
-                    await UpdateFileProgress(
-                            GameContextActionType.Download,
-                            accumulatedBytes,
-                            true,
-                            isPred
-                        )
-                        .ConfigureAwait(false);
-                }
-                if (totalWritten != chunkTotalSize)
-                {
-                    throw new IOException($"分片写入不完整: {totalWritten}/{chunkTotalSize}");
-                }
-                fileStream.SetLength(size);
-            }
-            catch (Exception ex)
-            {
-                Logger.WriteError($"下载文件{filePath}出现异常" + ex.Message);
-            }
-            finally
-            {
-                await fileStream.FlushAsync().ConfigureAwait(false);
-                await fileStream.DisposeAsync().ConfigureAwait(false);
-            }
-        }
     }
 
     private async Task<string?> DownloadFileByKrDiff(string dest, string filePath, bool isPred = false)
@@ -1784,135 +1134,8 @@ public partial class KuroGameContextBase
 
     #region 辅助方法
 
-    private bool IsDownloadCanceled()
-    {
-        return IsDownloadCanceled(false);
-    }
 
-    private bool IsDownloadCanceled(bool isPred)
-    {
-        var cts = isPred ? _prodDownloadCTS : _downloadCTS;
-        var state = isPred ? _prodDownloadState : _downloadState;
-        return cts == null || cts.IsCancellationRequested || state?.IsStop == true;
-    }
 
-    private CancellationTokenSource GetCTS(bool isPred)
-    {
-        return isPred ? (_prodDownloadCTS ?? _downloadCTS) : _downloadCTS;
-    }
-
-    private DownloadState GetState(bool isPred)
-    {
-        return isPred ? (_prodDownloadState ?? _downloadState) : _downloadState;
-    }
-
-    private async Task UpdateFileProgress(
-        GameContextActionType type,
-        long fileSize,
-        bool isAdd = true,
-        bool isPred = false,
-        string tip = ""
-    )
-    {
-        if (type == GameContextActionType.Download)
-        {
-            Interlocked.Add(ref _totalDownloadedBytes, fileSize);
-            if (isAdd)
-                Interlocked.Add(ref _totalProgressSize, fileSize);
-        }
-        else if (type == GameContextActionType.Verify)
-        {
-            if (!isAdd)
-                Interlocked.Add(ref _totalVerifiedBytes, fileSize);
-            if (isAdd)
-                Interlocked.Add(ref _totalProgressSize, fileSize);
-        }
-        var elapsed = (DateTime.Now - _lastSpeedUpdateTime).TotalSeconds;
-        if (elapsed >= 1)
-        {
-            _downloadSpeed = _totalDownloadedBytes / elapsed;
-            _verifySpeed = _totalVerifiedBytes / elapsed;
-            // 重置计数器和时间
-            Interlocked.Exchange(ref _totalDownloadedBytes, 0);
-            Interlocked.Exchange(ref _totalVerifiedBytes, 0);
-            var currentBytes = Interlocked.Read(ref _totalDownloadedBytes);
-            _lastSpeedBytes = currentBytes;
-            _lastSpeedUpdateTime = DateTime.Now;
-        }
-
-        var args = new GameContextOutputArgs
-        {
-            Type = type,
-            CurrentSize = _totalProgressSize,
-            TotalSize = _totalfileSize,
-            DownloadSpeed = _downloadSpeed,
-            VerifySpeed = _verifySpeed,
-            RemainingTime = RemainingTime,
-            IsAction = _downloadState?.IsActive ?? false,
-            IsPause = _downloadState?.IsPaused ?? false,
-            TipMessage = tip,
-        };
-        // keep separate last output args for predownload and normal download
-        if (isPred)
-            _lastProdOutputArgs = args;
-        else
-            _lastOutputArgs = args;
-
-        if (isPred && gameContextProdOutputDelegate != null)
-        {
-            await gameContextProdOutputDelegate.Invoke(this, args).ConfigureAwait(false);
-        }
-        else if (!isPred && gameContextOutputDelegate != null)
-        {
-            await gameContextOutputDelegate.Invoke(this, args).ConfigureAwait(false);
-        }
-    }
-
-    public TimeSpan RemainingTime
-    {
-        get
-        {
-            try
-            {
-                if (DownloadSpeed <= 0 || _totalDownloadedBytes >= _totalfileSize)
-                    return TimeSpan.Zero;
-
-                var remainingBytes = _totalfileSize - _totalProgressSize;
-                return TimeSpan.FromSeconds(remainingBytes / DownloadSpeed);
-            }
-            catch (Exception)
-            {
-                return TimeSpan.Zero;
-            }
-        }
-    }
-
-    public CDNSpeedTester CDNSpeedTester { get; private set; }
-
-    /// <summary>
-    /// 重新推送最后一次的下载/提示信息（用于页面切换后恢复显示）
-    /// </summary>
-    public async Task ReEmitLastOutputAsync(bool isPred = false)
-    {
-        if (isPred)
-        {
-            if (_lastProdOutputArgs == null)
-                return;
-            if (gameContextProdOutputDelegate != null)
-            {
-                await gameContextProdOutputDelegate.Invoke(this, _lastProdOutputArgs).ConfigureAwait(false);
-            }
-        }
-        else
-        {
-            if (_lastOutputArgs == null)
-                return;
-            if (gameContextOutputDelegate != null)
-            {
-                await gameContextOutputDelegate.Invoke(this, _lastOutputArgs).ConfigureAwait(false);
-            }
-        }
-    }
 
     #endregion
